Multi-GPU, allocate output tensor
Prior to this commit, `QuantLinear` and `QuantLlamaMLP` would allocate
their output tensor on `device="cuda"`.  This defaults to the first
GPU, equivalent to `"cuda:0"`, rather than the currently active GPU.
This commit updates the allocation of output tensors to instead occur
on `device=input.device`.

This is required for multi-GPU execution on systems with separate
memory spaces for each GPU (i.e. without NVLink).